Move code to convert.py
